% Auto-generated: do not edit by hand
\name{coastApp}

\alias{coastApp}

\title{CoastApp component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
coastApp(id=NULL, D0L_a2=NULL, E_L=NULL, Etrap=NULL, L_dist=NULL,
alpha=NULL, c0=NULL, c1=NULL, c2=NULL, c3=NULL, eta_q=NULL,
label=NULL, omega=NULL, psi=NULL, rad=NULL, rmr0=NULL,
th32=NULL, u38=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{D0L_a2}{Character. The value displayed in the input.}

\item{E_L}{Character. The value displayed in the input.}

\item{Etrap}{Character. The value displayed in the input.}

\item{L_dist}{Character. The value displayed in the input.}

\item{alpha}{Character. The value displayed in the input.}

\item{c0}{Character. The value displayed in the input.}

\item{c1}{Character. The value displayed in the input.}

\item{c2}{Character. The value displayed in the input.}

\item{c3}{Character. The value displayed in the input.}

\item{eta_q}{Character. The value displayed in the input.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{omega}{Character. The value displayed in the input.}

\item{psi}{Character. The value displayed in the input.}

\item{rad}{Character. The value displayed in the input.}

\item{rmr0}{Character. The value displayed in the input.}

\item{th32}{Character. The value displayed in the input.}

\item{u38}{Character. The value displayed in the input.}

\item{value}{Character. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

